refactor(ui-react): split useCopy hook into its own file
Keeps ClipboardProvider.tsx as a component-only module so
react-refresh/only-export-components no longer warns, and lets HMR
fast-refresh the provider cleanly.
