+++ ui-react/apps/console/src/components/common/ClipboardProvider.tsx
@@ -1,26 +1,7 @@
-import {
-  createContext,
-  ReactNode,
-  useCallback,
-  useContext,
-  useEffect,
-  useId,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import { ReactNode, useCallback, useId, useMemo, useState } from "react";
 import { ExclamationTriangleIcon } from "@heroicons/react/24/outline";
+import { ClipboardContext } from "@/hooks/useCopy";
 import BaseDialog from "./BaseDialog";
-
-// ─── Context ──────────────────────────────────────────────────────────────────
-
-interface ClipboardContextValue {
-  triggerWarning: () => void;
-}
-
-const ClipboardContext = createContext<ClipboardContextValue | null>(null);
-
-// ─── Provider ─────────────────────────────────────────────────────────────────
 
 /**
  * Mounts a single clipboard-warning dialog for the whole app.
@@ -80,60 +61,4 @@ export function ClipboardProvider({ children }: { children: ReactNode }) {
       </BaseDialog>
     </ClipboardContext.Provider>
   );
-}
-
-// ─── Hook ─────────────────────────────────────────────────────────────────────
-
-export interface UseCopyResult {
-  /** Call with the text to copy. Shows the warning dialog when clipboard access
-   *  is unavailable (insecure context or API error). */
-  copy: (text: string) => void;
-  /** True for 1500 ms after a successful copy. Use for inline visual feedback. */
-  copied: boolean;
-}
-
-/**
- * Safe clipboard copy with automatic insecure-context handling.
- *
- * Must be used within `<ClipboardProvider>`.
- *
- * ```tsx
- * const { copy, copied } = useCopy();
- * <button onClick={() => copy(deviceId)}>{copied ? "Copied!" : "Copy"}</button>
- * ```
- */
-export function useCopy(): UseCopyResult {
-  const ctx = useContext(ClipboardContext);
-  if (!ctx) throw new Error("useCopy must be used within <ClipboardProvider>");
-
-  const { triggerWarning } = ctx;
-  const [copied, setCopied] = useState(false);
-  const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  const copy = useCallback(
-    (text: string) => {
-      if (!globalThis.isSecureContext) {
-        triggerWarning();
-        return;
-      }
-
-      navigator.clipboard.writeText(text).then(
-        () => {
-          if (timerRef.current) clearTimeout(timerRef.current);
-          setCopied(true);
-          timerRef.current = setTimeout(() => setCopied(false), 1500);
-        },
-        () => triggerWarning(),
-      );
-    },
-    [triggerWarning],
-  );
-
-  useEffect(() => {
-    return () => {
-      if (timerRef.current) clearTimeout(timerRef.current);
-    };
-  }, []);
-
-  return { copy, copied };
 }
